requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
pytest==8.4.2
httpx==0.28.1
pyyaml==6.0.3